fix: move static functions of ItineraryPage into a separate file 1;95;0c To split the huge component to manageable parts
